fix: migrate executeAction code from Editor to EditorLauncher
