+++ src/components/EditorLauncher/EditorLauncher.tsx
@@ -5,17 +5,40 @@ import { StartIcon, StopIcon } from "../Icons";
 import * as monaco from "monaco-editor-core";
 import { useEditorContext } from "../EditorProvider";
 import LoadingSpinner from "../LoadingSpinner";
+import { useEffectOnce } from "@/util/useEffectOnce";
 
 const EditorLauncher = ({ onStartQuery, onStopQuery, ...props }: EditorLauncherProps) => {
   const { editor, isQueryStarting, isQueryStopping } = useEditorContext();
+
+  useEffectOnce(() => {
+    if (!editor) return;
+
+    const runAction: monaco.editor.IActionDescriptor = {
+      id: "runSelectedOrEntireQuery",
+      label: "Run (selected) query",
+      contextMenuOrder: 2,
+      contextMenuGroupId: "1_modification",
+      keybindings: [monaco.KeyMod.CtrlCmd | monaco.KeyCode.Enter],
+      run: editor => {
+        const model = editor.getModel();
+        // by default, only run selected value
+        let selectedValue = model.getValueInRange(editor.getSelection());
+        // however, if nothing is selected, we run all editor value
+        if (selectedValue === "") {
+          selectedValue = model.getValue();
+        }
+
+        onStartQuery && onStartQuery(editor, selectedValue);
+      },
+    };
+
+    editor.addAction(runAction);
+  });
+
   return (
     <div css={EditorLauncherStyle} {...props}>
-      {isQueryStopping ? <LoadingSpinner /> : <StopIcon onClick={() => !isQueryStarting && !isQueryStopping && onStopQuery && onStopQuery(editor)} />}
-      {isQueryStarting ? (
-        <LoadingSpinner />
-      ) : (
-        <StartIcon onClick={() => !isQueryStarting && !isQueryStopping && editor.trigger("", "executeSelectedCode", "")} />
-      )}
+      {isQueryStopping ? <LoadingSpinner /> : <StopIcon onClick={() => !isQueryStarting && onStopQuery && onStopQuery(editor)} />}
+      {isQueryStarting ? <LoadingSpinner /> : <StartIcon onClick={() => !isQueryStopping && editor.trigger("", "runSelectedOrEntireQuery", "")} />}
     </div>
   );
 };
@@ -33,7 +56,7 @@ export interface EditorLauncherProps extends HTMLAttributes<HTMLDivElement> {
 }
 
 export type EditorLauncherEventHandler = {
-  method(editor?: monaco.editor.IStandaloneCodeEditor, targetValue?: string): void;
+  method(editor?: monaco.editor.ICodeEditor, targetValue?: string): void;
 }["method"];
 
 export default EditorLauncher;
